backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart
pydantic>=2.0.0
python-jose
httpx==0.27.0
rich
python-dotenv
werkzeug>=3.0.0
cryptography>=41.0.0
requests
pynacl
py-multibase
shamirs
